+++ programs/unstake/src/instructions/add_liquidity.rs
@@ -2,7 +2,11 @@ use anchor_lang::{prelude::*, system_program};
 use anchor_spl::token::{self, Mint, MintTo, Token, TokenAccount};
 use std::convert::TryFrom;
 
-use crate::{errors::UnstakeError, state::Pool};
+use crate::{
+    errors::UnstakeError,
+    state::{Pool, FLASH_ACCOUNT_SEED_SUFFIX},
+    utils::calc_pool_owned_lamports,
+};
 
 #[derive(Accounts)]
 pub struct AddLiquidity<'info> {
@@ -36,6 +40,15 @@ pub struct AddLiquidity<'info> {
     )]
     pub mint_lp_tokens_to: Account<'info, TokenAccount>,
 
+    /// amount taken for all active flash loans of the pool
+    /// CHECK: PDA checked
+    #[account(
+        mut,
+        seeds = [&pool_account.key().to_bytes(), FLASH_ACCOUNT_SEED_SUFFIX],
+        bump,
+    )]
+    pub flash_account: UncheckedAccount<'info>,
+
     pub token_program: Program<'info, Token>,
     pub system_program: Program<'info, System>,
 }
@@ -48,14 +61,13 @@ impl<'info> AddLiquidity<'info> {
         let pool_sol_reserves = &ctx.accounts.pool_sol_reserves;
         let lp_mint = &ctx.accounts.lp_mint;
         let mint_lp_tokens_to = &ctx.accounts.mint_lp_tokens_to;
+        let flash_account = &ctx.accounts.flash_account;
         let token_program = &ctx.accounts.token_program;
         let system_program = &ctx.accounts.system_program;
 
         // order matters, must calculate first before mutation
-        let pool_owned_lamports = pool_sol_reserves
-            .lamports()
-            .checked_add(pool_account.incoming_stake)
-            .ok_or(UnstakeError::InternalError)?;
+        let pool_owned_lamports =
+            calc_pool_owned_lamports(pool_sol_reserves, pool_account, flash_account)?;
         let to_mint = calc_lp_tokens_to_mint(pool_owned_lamports, lp_mint.supply, amount)?;
 
         // transfer SOL