+++ programs/unstake/src/rational.rs
@@ -1,10 +1,11 @@
 use anchor_lang::prelude::*;
+use serde::Deserialize;
 use spl_math::precise_number::PreciseNumber;
 
 use std::{convert::TryInto, fmt};
 
 /// A ratio. Denom should not = 0
-#[derive(Debug, PartialEq, Clone, Copy, AnchorSerialize, AnchorDeserialize)]
+#[derive(Debug, PartialEq, Clone, Copy, AnchorSerialize, AnchorDeserialize, Deserialize)]
 pub struct Rational {
     pub num: u64,
     pub denom: u64,