+++ ts/src/index.ts
@@ -4,8 +4,7 @@ export * from "@project-serum/anchor";
 
 // The types for the JSON file and the IDL const is different
 // so we need to re-export both
-import * as IDL_JSON from "./idl/idl.json";
-export { IDL_JSON };
+export { IDL_JSON } from "./idl/idlJson";
 export * from "./idl/idl";
 
 export * from "./fee";